+++ assets/js/csp/modules/schoolSetup/Configuration.js
@@ -15,6 +15,16 @@ csp.modules.schoolSetup.configuration = {
 			$(el).css('visibility', (i <= score ? 'visible' : 'hidden'));
 		});
 	},
+	// @since 13.0 Add "Credit Food Service Account on Lunch Payment" config option
+	studentBillingFoodServiceAccountToggle: function() {
+		$('.student-billing-credit-food-service-account-options').toggleClass(
+			'hide',
+			! this.checked
+		).attr(
+			'disabled',
+			! this.checked
+		);
+	},
 	ready: function() {
 		var $input = $('input[name="values[config][PASSWORD_STRENGTH]"');
 
@@ -22,6 +32,10 @@ csp.modules.schoolSetup.configuration = {
 		csp.modules.schoolSetup.configuration.passwordStrengthBarsScore.call($input[0]);
 
 		$input.on('change', csp.modules.schoolSetup.configuration.passwordStrengthBarsScore);
+
+		var $input2 = $('.onclick-student-billing-credit-food-service-account-toggle');
+
+		$input2.on('change', csp.modules.schoolSetup.configuration.studentBillingFoodServiceAccountToggle);
 	}
 }